svn merge -r 29043:29044 svn://chef/work/src . (Comment: #8788)
git-svn-id: https://ut.svn.beanstalkapp.com/ngfw/branch/prod/release-9.0/work/src@29045 8ebf8b4e-06f3-0310-961f-89f74618be54

+++ cpd/api/com/untangle/node/cpd/BlockEvent.java
@@ -30,7 +30,7 @@ public class BlockEvent extends LogEvent implements Serializable
 
     private short protocol;
 
-    private byte clientIntf;
+    private int clientIntf;
 
     private InetAddress clientAddress;
     private InetAddress serverAddress;
@@ -79,12 +79,12 @@ public class BlockEvent extends LogEvent implements Serializable
      * @return the number of the interface of the client
      */
     @Column(name="client_intf", nullable=false)
-    public byte getClientIntf()
+    public int getClientIntf()
     {
         return clientIntf;
     }
 
-    public void setClientIntf(byte clientIntf)
+    public void setClientIntf(int clientIntf)
     {
         this.clientIntf = clientIntf;
     }
